refactor useAssignments
