+++ pages/user.tsx
@@ -11,7 +11,6 @@ import { useFilterContext } from "../src/context/filter-context";
 import { useEffect, useMemo, useState } from "react";
 import { ContentHeader } from "../src/components/content-header";
 import { useStats } from "../src/hooks/use-stats";
-import { useAssignments } from "../src/hooks/use-assignments";
 import { useHours } from "../src/hooks/use-hours";
 import dynamic from "next/dynamic";
 import { PieChartProps } from "reaviz/dist/src/PieChart/PieChart";
@@ -35,7 +34,7 @@ import { getColor } from "../src/utils/get-color";
 import { AreasChart } from "../src/components/chart/areas-chart";
 import { DATE_FORMAT } from "../src/context/formats";
 import { withServerSideSession } from "../src/server/with-session";
-import {useEntries} from "../src/hooks/use-remote";
+import {DefaultParams, useAssignments, useEntries} from "../src/hooks/use-remote";
 
 //@ts-ignore
 const PieChart = dynamic<PieChartProps>(() => import('reaviz').then(module => module.PieChart), { ssr: false });
@@ -103,15 +102,16 @@ export const User = ({
     const from = format(dateRange[0] ?? new Date(), DATE_FORMAT)
     const to = format(dateRange[1] ?? new Date(), DATE_FORMAT)
 
-    const apiParams = {
+    const apiParams : DefaultParams = {
         from, to, uid: userId
     }
 
     const entriesApi = useEntries(apiParams);
+    const assignmentsApi = useAssignments(apiParams);
+
     const entriesDetailedApi = useEntriesDetailed();
     const currentStatsApi = useCurrentStats();
     const statsApi = useStats();
-    const assignmentsApi = useAssignments();
     const hoursApi = useHours();
 
     const amountOfDays = useMemo(() => differenceInBusinessDays(dateRange[1], dateRange[0]) + 1, [ dateRange ]);
@@ -122,7 +122,6 @@ export const User = ({
         }
         entriesDetailedApi.load(from, to, userId);
         statsApi.load(from, to, userId);
-        assignmentsApi.load(from, to, userId);
         hoursApi.load(from, to, userId);
         currentStatsApi.load(userId);
 
@@ -294,7 +293,7 @@ export const User = ({
                                                 padRadius={ 200 }
                                                 doughnut={ true }
                                             /> }
-                                            data={ (assignmentsApi.assignments ?? []).map((h) => ({
+                                            data={ (assignmentsApi.data?.assignments ?? []).map((h) => ({
                                                 key: h.name ?? h.code ?? '?',
                                                 data: h.totalHours ?? 0
                                             })) ?? [] }/>
